backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
requests
elevenlabs
pytest==8.0.0
httpx==0.26.0 